+++ utilities/router.go
@@ -10,10 +10,11 @@ import (
 type Router struct {
 	userHandler    *handler.UserHandler
 	companyHandler *handler.CompanyHandler
+	towHandler     *handler.TowHandler
 }
 
-func NewRouter(user *handler.UserHandler, company *handler.CompanyHandler) *Router {
-	return &Router{userHandler: user, companyHandler: company}
+func NewRouter(user *handler.UserHandler, company *handler.CompanyHandler, towHandler *handler.TowHandler) *Router {
+	return &Router{userHandler: user, companyHandler: company, towHandler: towHandler}
 }
 
 // InitializeRouter builds the gin.Engine and registers routes/middleware.
@@ -43,6 +44,9 @@ func (r *Router) InitializeRouter() *gin.Engine {
 	// ==== Company routes ====
 	engine.POST("/company", r.companyHandler.PostCompany)   // Create a company
 	engine.GET("/company/:id", r.companyHandler.GetCompany) // Get a company
+
+	// ==== Tow routes ====
+	engine.GET("/tows/company/:companyId", r.towHandler.GetTowHistory) // Get a company
 
 	return engine
 }